eval/inference/RadFM/requirements.txt
einops==0.8.0
einops_exts==0.0.4
numpy==1.24.1
numpy==1.24.3
pandas==2.0.1
Pillow==9.4.0
Pillow==11.0.0
scipy==1.14.1
timm==0.6.13
torch==2.4.1+cu118
torch==1.13.0
torchscale.egg==info
torchvision==0.19.1+cu118
torchvision==0.14.0
tqdm==4.67.1
transformers==4.28.1
